Update primary brand color
