+++ web/src/pages/filter/index.tsx
@@ -315,7 +315,7 @@ function MineFilter() {
     () => (
       <StarsBackground
         pointerEvents={false}
-        starColor="#07C160"
+        starColor="#3ECF4A"
         className="relative h-auto max-h-[25vh] overflow-hidden bg-none"
       >
         <div className="noScrollBar relative max-h-[25vh] space-y-4 overflow-y-scroll bg-none mask-[linear-gradient(180deg,#000000_calc(100%-20%),transparent)] p-4 font-semibold">
