+++ web/src/styles/tailwindTheme.css
@@ -4,12 +4,12 @@
     /* 全局缩放倍数 - 可以调整这个值 */
 
     /* Background Images */
-    --background-image-primary: linear-gradient(272deg, rgba(7, 193, 96, 0.96) -0.55%, rgba(178, 207, 62, 0.96) 99.45%);
-    --background-image-primaryGreenGradient: linear-gradient(327deg, #07C160 18.08%, rgba(62, 207, 74, 0.96) 64.28%, rgba(153, 230, 38, 0.40) 111.94%, rgba(250, 255, 0, 0.00) 158.87%);
+    --background-image-primary: linear-gradient(272deg, rgba(62, 207, 74, 0.96) -0.55%, rgba(178, 207, 62, 0.96) 99.45%);
+    --background-image-primaryGreenGradient: linear-gradient(327deg, #3ECF4A 18.08%, rgba(62, 207, 74, 0.96) 64.28%, rgba(153, 230, 38, 0.40) 111.94%, rgba(250, 255, 0, 0.00) 158.87%);
     --background-image-pattern: repeating-linear-gradient(315deg, var(--pattern-fg) 0, var(--pattern-fg) 1px, transparent 0, transparent 50%);
 
     /* Colors */
-    --color-theme: #07C160;
+    --color-theme: #3ECF4A;
 
     /* Animations */
     --animate-meteor-effect: meteor 5s linear infinite;
